Move application/errors to application/http-errors
