+++ api/lib/application/competence-evaluations/competence-evaluation-controller.js
@@ -1,7 +1,7 @@
+const { BadRequestError } = require('../http-errors');
 const usecases = require('../../domain/usecases');
 const serializer = require('../../infrastructure/serializers/jsonapi/competence-evaluation-serializer');
 const queryParamsUtils = require('../../infrastructure/utils/query-params-utils');
-const { BadRequestError } = require('../../application/errors');
 
 module.exports = {
