+++ api/lib/application/feedbacks/feedback-controller.js
@@ -1,7 +1,7 @@
+const { BadRequestError } = require('../http-errors');
 const _ = require('../../infrastructure/utils/lodash-utils');
 const serializer = require('../../infrastructure/serializers/jsonapi/feedback-serializer');
 const repository = require('../../infrastructure/repositories/feedback-repository');
-const { BadRequestError } = require('../../application/errors');
 
 module.exports = {
