+++ api/lib/infrastructure/utils/ods/read-ods-utils.js
@@ -1,7 +1,8 @@
-const { UnprocessableEntityError } = require('../../../application/errors');
-const { loadOdsZip } = require('./common-ods-utils');
 const _ = require('lodash');
 const XLSX = require('xlsx');
+
+const { UnprocessableEntityError } = require('../../../application/http-errors');
+const { loadOdsZip } = require('./common-ods-utils');
 
 const CONTENT_XML_IN_ODS = 'content.xml';
