+++ api/tests/test-helper.js
@@ -1,5 +1,5 @@
 // Errors
-const infraErrors = require('../lib/application/errors');
+const httpErrors = require('../lib/application/http-errors');
 // Chai
 const chai = require('chai');
 const expect = chai.expect;
@@ -181,6 +181,6 @@ module.exports = {
   streamToPromise,
   catchErr,
   testErr: new Error('Fake Error'),
-  testInfraNotFoundErr: new infraErrors.NotFoundError('Fake infra NotFoundError'),
+  testInfraNotFoundErr: new httpErrors.NotFoundError('Fake infra NotFoundError'),
   compareDatabaseObject
 };
